requirements.txt
# ============================================================
# requirements.txt
# ============================================================

--extra-index-url https://download.pytorch.org/whl/cu121

#Web Framework
django>=4.2,<5.0
djangorestframework>=3.14

# Network packet capture and parsing
scapy>=2.5

#Data processing
numpy>=1.26.4
pandas>=2.0

# image processing
Pillow>=10.0

# Deep Learning (GPU Version)
torch==2.1.0+cu121
torchvision==0.16.0+cu121

# Machine learning tools
scikit-learn>=1.3

# Data Visualization
matplotlib>=3.7
plotly>=5.0

# Terminal Tools
colorama>=0.4.6
tabulate>=0.9.0

# HTTP Request (for AI Agent n8n concatenation)
requests>=2.31

# Test framework
pytest>=7.0
pytest-cov>=4.0
pytest-django>=4.5
